requirements.txt
-e git://github.com/smiley/steamapi.git#egg=steamapi
python-slugify
